requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.7
django-admin-charts==0.11.0
django-admin-tools==0.8.1
django-admin-tools-stats==0.9.0
django-allauth==0.40.0
django-bootstrap3==11.1.0
django-bower==5.2.0
django-dropbox-storage==0.4.1
django-heroku==0.3.1
django-import-export==1.2.0
django-jsonfield==1.3.1
django-material-admin==1.6.7
django-nvd3==0.9.7
django-qsstats-magic==1.1.0
djcacheutils==3.0.0
dropbox==9.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
Jinja2==2.10.3
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
psycopg2==2.7.5
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
Unidecode==1.1.1
urllib3==1.25.7
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
